feat: Update page title and favicon
The code changes update the page title and favicon in the navbar.html template. The page title is changed to "Gen-Zote | Missing Person Database" and the favicon is updated to use the "/static/images/favicon.jpg" file. This will provide a more accurate and visually appealing representation of the website.